update doc comments for UintRef bits methods

+++ src/uint/ref_type/bits.rs
@@ -3,11 +3,13 @@ use crate::{ConstChoice, Limb, traits::BitOps};
 use subtle::Choice;
 
 impl UintRef {
-    /// Get the precision of this [`BoxedUint`] in bits.
+    /// Get the precision of this number in bits.
     pub const fn bits_precision(&self) -> u32 {
         self.0.len() as u32 * Limb::BITS
     }
 
+    /// Get the value of the bit at position `index`, as a truthy or falsy [`ConstChoice`].
+    /// Returns the falsy value for indices out of range.
     #[inline(always)]
     pub const fn bit(&self, index: u32) -> ConstChoice {
         let limb_num = index / Limb::BITS;
@@ -26,10 +28,15 @@ impl UintRef {
         ConstChoice::from_word_lsb(result >> index_in_limb)
     }
 
+    /// Returns `true` if the bit at position `index` is set, `false` for an unset bit
+    /// or for indices out of range.
+    ///
+    /// # Remarks
+    /// This operation is variable time with respect to `index` only.
     #[inline(always)]
     pub const fn bit_vartime(&self, index: u32) -> bool {
         let limb_num = (index / Limb::BITS) as usize;
-        let index_in_limb = (index % Limb::BITS) as usize;
+        let index_in_limb = index % Limb::BITS;
         if limb_num >= self.0.len() {
             false
         } else {
@@ -47,6 +54,8 @@ impl UintRef {
         self.bits_precision() - self.leading_zeros()
     }
 
+    /// Calculate the number of bits needed to represent this number in variable-time with respect
+    /// to `self`.
     #[inline(always)]
     pub const fn bits_vartime(&self) -> u32 {
         let mut i = self.0.len() - 1;
@@ -58,6 +67,7 @@ impl UintRef {
         Limb::BITS * (i as u32 + 1) - limb.leading_zeros()
     }
 
+    /// Sets the bit at `index` to 0 or 1 depending on the value of `bit_value`.
     #[inline(always)]
     pub const fn set_bit(&mut self, index: u32, bit_value: ConstChoice) {
         let limb_num = index / Limb::BITS;
@@ -74,6 +84,8 @@ impl UintRef {
         }
     }
 
+    /// Sets the bit at `index` to 0 or 1 depending on the value of `bit_value`, in variable-time
+    /// with respect to `index`.
     #[inline(always)]
     pub const fn set_bit_vartime(&mut self, index: u32, bit_value: bool) {
         let limb_num = (index / Limb::BITS) as usize;
@@ -102,6 +114,7 @@ impl UintRef {
         count
     }
 
+    /// Calculate the number of trailing zeros in the binary representation of this number.
     #[inline(always)]
     pub const fn trailing_zeros(&self) -> u32 {
         let mut count = 0;
@@ -119,6 +132,8 @@ impl UintRef {
         count
     }
 
+    /// Calculate the number of trailing zeros in the binary representation of this number, in
+    /// variable-time with respect to `self`.
     #[inline(always)]
     pub const fn trailing_zeros_vartime(&self) -> u32 {
         let mut count = 0;
@@ -136,6 +151,7 @@ impl UintRef {
         count
     }
 
+    /// Calculate the number of trailing ones in the binary representation of this number.
     #[inline(always)]
     pub const fn trailing_ones(&self) -> u32 {
         let mut count = 0;
@@ -153,6 +169,8 @@ impl UintRef {
         count
     }
 
+    /// Calculate the number of trailing ones in the binary representation of this number, in
+    /// variable-time with respect to `self`.
     #[inline(always)]
     pub const fn trailing_ones_vartime(&self) -> u32 {
         let mut count = 0;
